Fix incorrect array default values (#37)
### Changelog
Fix incorrect serialization when values for array fields are not
specified

### Docs
None

### Description
When a message that is to be serialized contains less fields than the
given schema, the MessageWriter will take the default value for the
missing fields. For arrays of non-fixed length, the default value is an
empty array. However, prior to this PR, the default value for these
arrays was not `[]` but an array with a single element. This was caused
by having the array length being undefined in the Array constructor:

```
new Array(undefined).fill(0)
-> [0]
```

Changing this as shown below fixes the issue:
```
new Array(undefined ?? 0).fill(0)
-> []
```